+++ ec-customer-srv/src/main/java/com/ec/customer/repository/PermissionRepository.java
@@ -3,5 +3,5 @@ package com.ec.customer.repository;
 import com.ec.customer.model.entity.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, Integer> {
 }
